requirements.txt
flask==3.0.2
numpy==1.26.4
pillow==10.2.0
opencv-python-headless==4.9.0.80  
matplotlib==3.8.3
scipy==1.13.0